tests/wpt/web-platform-tests/bluetooth/requestDevice/name-empty-device-from-name-empty-filter.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'An empty name device can be obtained by empty name filter.'

bluetooth_test(
    () =>
        setUpPreconnectedDevice({name: ''})
            .then(() => requestDeviceWithTrustedClick({filters: [{name: ''}]}))
            .then(device => assert_equals(device.name, '')),
    test_desc);
</script>
